==> 01_PreparingDataTraining.ipynb <==
import pandas as pd
import numpy as np
from category_encoders import TargetEncoder
from sklearn.model_selection import train_test_split
from xgboost import XGBClassifier
import pickle

##### Leer Dataset de Entrenamiento

# Leer el dataset
df = pd.read_csv("datasets/MasterBI_Train.csv", header=0, sep=",", encoding="utf-8")
# Crear una serie con la variable target
y = df['HasDetections']

# Mostrar todas las columnas del dataframe
pd.set_option('display.max_columns',None)
df.head()


# Definir las variables identifier y target
identifier = 'MachineIdentifier' #no aporta valor al modelo
target = 'HasDetections' #variable objetivo

# Eliminar las columnas 'HasDetections' y 'MachineIdentifier' del dataframe
df.drop('HasDetections', axis=1, inplace=True)
df.drop('MachineIdentifier', axis=1, inplace=True)

# Llenar los Nulls de la variable 'Census_IsWIMBootEnabled' con 1 
df['Census_IsWIMBootEnabled'] = df['Census_IsWIMBootEnabled'].fillna(1)

# Llenar los Nulls de la variable 'PuaMode' con 1 
df['PuaMode'] = df['PuaMode'].fillna('off')
df['PuaMode'] = df['PuaMode'].replace({'off': 0, 'on': 1})

##### Eliminar variables con Alta Cardinalidad y con demasiados valores Nulos

# 1. Variables con alta cardinalidad
high_cardinality_threshold = 0.10  # 10% del total de filas
high_cardinality = []
for col in df.columns:
    n_unique = df[col].nunique(dropna=False)
    if n_unique / df.shape[0] > high_cardinality_threshold:
        high_cardinality.append(col)

# 2. Variables con demasiados valores desconocidos
missing_values_threshold = 0.60  # 30%
missing_values = []
for col in df.columns:
    if df[col].isna().sum() / df.shape[0] > missing_values_threshold:
        missing_values.append(col)

# Eliminar columnas con Alta cardinalidad
df.drop(high_cardinality, axis=1, inplace=True)

# Eliminar columnas con demasiados valores desconocidos
df.drop(missing_values, axis=1, inplace=True)

features = [col for col in df.columns if col not in [identifier, target]]

##### Clasificar las variables en Categoricas, Númericas y Booleanas

# Clasificación de variables
def clasificar_variables(df, cols):
    binary_cols, cat_cols, num_cols = [], [], []
    for col in cols:
        n_classes = df[col].nunique()
        if n_classes <= 2:
            binary_cols.append(col)
        elif df.shape[0] * 0.01 > n_classes > 2:
            cat_cols.append(col)
        else:
            num_cols.append(col)
    return binary_cols, cat_cols, num_cols

# Clasificar las variables
binary_cols, cat_cols, num_cols = clasificar_variables(df, features)

# Guardar en un diccionario las variables clasificadas
clasificacion_vars = {
    'binary_cols': binary_cols,
    'cat_cols': cat_cols,
    'num_cols': num_cols
}

# Guardar en archivo pickle el diccionario de variables clasificadas
with open("files/clasificacion_vars.pkl", "wb") as f:
    pickle.dump(clasificacion_vars, f)



##### Imputar valores nulos en dependencia al tipo de variable. Luego guardar los datos que se imputaron para cada columna.

# Diccionario para guardar los valores de imputación
valores_imputacion = {
    'binarias': {},
    'categoricas': {},
    'numericas': {}
}

# Imputación de variables binarias
def imputar_binarias(df, cols):
    for col in cols:
        if df[col].isnull().sum() == 0:
            continue
        value_counts = df[col].value_counts(normalize=True)
        if len(value_counts) == 0 or value_counts.max() > 0.8:
            df[col] = df[col].fillna(-1)
            valores_imputacion['binarias'][col] = -1
        elif abs(value_counts.get(0, 0) - value_counts.get(1, 0)) < 0.1:
            moda = df[col].mode()[0]
            df[col] = df[col].fillna(moda)
            valores_imputacion['binarias'][col] = moda
        else:
            probs = value_counts.to_dict()
            df[col] = df[col].apply(lambda x: np.random.choice([0, 1], p=[probs.get(0, 0.5), probs.get(1, 0.5)]) if pd.isnull(x) else x)
            valores_imputacion['binarias'][col] = probs
    return df

# Imputación de variables categóricas
def imputar_categoricas(df, cols):
    for col in cols:
        if df[col].isnull().sum() == 0:
            continue
        moda = df[col].mode()[0]
        freq = df[col].value_counts(normalize=True).get(moda, 0)
        valor = "missing" if freq > 0.8 else moda
        df[col] = df[col].fillna(valor)
        valores_imputacion['categoricas'][col] = valor
    return df

# Imputación de variables numéricas
def imputar_numericas(df, cols):
    for col in cols:
        if df[col].dtype == 'object':       
            try:
                df[col] = df[col].astype('float64')
            except ValueError:
                continue

        if df[col].isnull().sum() == 0:
            continue
        skew = df[col].dropna().skew()
        valor = df[col].mean() if abs(skew) < 0.5 else df[col].median()
        df[col] = df[col].fillna(valor)
        valores_imputacion['numericas'][col] = valor
    return df

# Aplicar imputaciones
df = imputar_binarias(df, binary_cols)
df = imputar_categoricas(df, cat_cols)
df = imputar_numericas(df, num_cols)

# Guardar el diccionario con lo valores de imputación por columna
with open("files/valores_imputacion.pkl", "wb") as f:
    pickle.dump(valores_imputacion, f)


##### Aplicar Target Encoding a las variables categoricas y guardar el encoder

# Aplicar el Target Encoding
X = df.copy()
te = TargetEncoder(cols = cat_cols, smoothing = 10)
te.fit(X[cat_cols], y)
X_te= X.join(te.transform(X[cat_cols]).add_suffix('_target'))
X_te = X_te.drop(cat_cols, axis=1)

# Guardar el TargetEncoder en un archivo
with open("files/target_encoder.pkl", "wb") as archivo:
    pickle.dump(te, archivo)



X = X_te.copy()
# Listar el nuevo nombre de las variables
features = [col for col in X.columns if col not in [identifier, target]]

# Dividir el dataset en entrenamiento y prueba (20% prueba, 80% entrenamiento)
X_train, X_test, y_train, y_test =  train_test_split(X, y, test_size = 0.2, random_state = 42)

##### Entrenar un modelo inicial para determinar las variables con mayor importancia a la hora de predecir

# Entrenar un modelo inicial con el clasificador XGBoost
xgb_model = XGBClassifier()
xgb_model.fit(X_train, y_train)

# Realizar predicciones en el conjunto de prueba
xgb_predict = xgb_model.predict(X_test)

# Crear un DataFrame para la importancia de las variables
xgb_features = pd.DataFrame(features)
# Importancia de las variables según el modelo XGBoost
xgb_features['score'] = xgb_model.feature_importances_
xgb_features.columns = ['features','feature_score']
# Ordenar las variables por importancia
xgb_features.sort_values(by=['feature_score'],ascending = False,inplace = True)
# Mostrar las 20 variables más importantes
xgb_features.head(20)

# Seleccionar/Filtrar las variables con importancia mayor a 0.01
xgb_features_importants = xgb_features[xgb_features['feature_score'] > 0.01]
# Guardar la lista de variables más importantes
final_features = xgb_features_importants.features.to_list()

# Guardar en archivo pickle la lista de variables finales a usar
with open("files/final_features.pkl", "wb") as f:
    pickle.dump(final_features, f)

##### Guardar los nuevos datos de entrenamiento para usarlos en las siguientes etapas

X = X.loc[:, final_features]
y = y.copy()

datos_entrenamiento = {
    'X': X,
    'y': y
}

# Guardar en archivo pickle
with open("files/datos_entrenamiento.pkl", "wb") as f:
    pickle.dump(datos_entrenamiento, f)
