==> 02_SearchHyperparameters.ipynb <==
import pickle
from sklearn.model_selection import train_test_split

from lightgbm import LGBMClassifier
import numpy as np

import xgboost as xgb
from sklearn.model_selection import RandomizedSearchCV

##### Leer los datos de entrenamiento limpios previamente guardados

# Leer los datos de entrenamiento desde el archivo pickle
with open("files/datos_entrenamiento.pkl", "rb") as archivo:
    datos_entrenamiento = pickle.load(archivo)

X = datos_entrenamiento['X']
y = datos_entrenamiento['y']

# Dividir el dataset en entrenamiento y prueba (20% prueba, 80% entrenamiento)
X_train,X_test,y_train,y_test = train_test_split(X, y, test_size = 0.2,random_state = 42)

##### Buscar los mejores Hiperparámetros para un modelo XGB

# Distribución de los hiperparámetros a ajustar para XGBoost
param_dist = {
        'n_estimators' : [200,250,300,350,400], # no of trres
        'learning_rate' : [0.01],
        'min_child_weight': [4],
        'gamma': [0.5,0.7,1.0,1.3],
        'colsample_bytree': [0.5],
        'subsample' : [0.3],
        'max_depth': [10],
        'learning_rate' : [0.01], 
        'reg_lambda' : [0.6],
        'objective' :['binary:logistic'],
        'eval_metric':['auc']
        }

#folds = 5
#param_comb = 500

# Modelo base XGBoost     
xgb_model = xgb.XGBClassifier(n_jobs = -1)

# Validación cruzada
cv = 3

random_search_xgb = RandomizedSearchCV(
        estimator=xgb_model, 
        param_distributions=param_dist, 
        n_iter=500, 
        scoring='roc_auc', 
        refit = 'roc_auc',
        n_jobs=-1, 
        cv=cv, 
        verbose=3, 
        random_state=42)

random_search_xgb.fit(X_train, y_train)

# Mostrar los mejores hiperparámetros encontrados para XGBoost
print(random_search_xgb.best_params_)

# Mostrar la mejor puntuación obtenida con esos hiperparámetros para XGBoost
print(random_search_xgb.best_score_)

# Mostrar el mejor estimador
print(random_search_xgb.best_estimator_)

##### Buscar los mejores Hiperparámetros para un modelo LGBM

# Distribución de los hiperparámetros a ajustar para LGBMClassifier
param_dist = {
            'n_estimators' : np.arange(20,100,5),
            'feature_fraction' : np.arange(0.1,1,0.1),
            'bagging_fraction' : np.arange(0.5,0.99,0.05),
            'num_leaves' : np.arange(30,100,10),
            'learning_rate' : np.arange(0.01,0.5,0.05),
            'max_depth' : np.arange(10,100,5),
            'min_child_samples' : np.arange(2,7,1),
            'reg_alpha' : np.arange(0,1,0.2),
            'reg_lambda' : np.arange(0,1,0.2),
            'colsample_bytree' : np.arange(0,1,0.2)
}

# Modelo base LGBM
lgbm = LGBMClassifier(random_state=42, n_jobs=1)

# Validación cruzada
cv = 3 

# RandomizedSearchCV
random_search_lgbm = RandomizedSearchCV(
    estimator=lgbm,
    param_distributions=param_dist,
    n_iter=50,
    scoring='roc_auc',
    cv=cv,
    verbose=1,
    random_state=42,
    n_jobs=-1
)

random_search_lgbm.fit(X_train, y_train)

# Mostrar los mejores hiperparámetros encontrados para LGBMClassifier
print(random_search_lgbm.best_params_)

# Mostrar la mejor puntuación obtenida con esos hiperparámetros para LGBMClassifier
print(random_search_lgbm.best_score_)

# Mostrar el mejor estimador para LGBMClassifier
print(random_search_lgbm.best_estimator_)
