==> 05_ExecuteModels.ipynb <==
# Importar las librerías necesarias
import pickle

# Cargar los datos de test desde el archivo pickle
with open("files/datos_test.pkl", "rb") as f:
    datos_test = pickle.load(f)

X_test = datos_test['X']
kaggle_submission_xgb = datos_test['kaggle_submission']
kaggle_submission_lgbm = datos_test['kaggle_submission']

# Cargar el modelo XGB entrenado desde el archivo pickle
with open("files/model_xgb.pkl", "rb") as f:
    model_xgb = pickle.load(f)

# Obtener predicciones (0 o 1)
predicciones = model_xgb.predict(X_test)
# Probabilidades (entre 0 y 1)
probabilidades = model_xgb.predict_proba(X_test)[:, 1]

kaggle_submission_xgb["HasDetections"] = probabilidades
kaggle_submission_xgb.to_csv("datasets/predicciones_modelo_xgb.csv", index=False)

# Cargar el modelo XGB entrenado desde el archivo pickle
with open("files/model_lgbm.pkl", "rb") as f:
    model_lgbm = pickle.load(f)

# Obtener predicciones (0 o 1)
predicciones = model_lgbm.predict(X_test)
# Probabilidades (entre 0 y 1)
probabilidades = model_lgbm.predict_proba(X_test)[:, 1]

kaggle_submission_lgbm["HasDetections"] = probabilidades
kaggle_submission_lgbm.to_csv("datasets/predicciones_modelo_lgbm.csv", index=False)
